+++ lib/libesp32/Berry/default/be_modtab.c
@@ -132,6 +132,10 @@ extern void be_load_AXP192_class(bvm *vm);
 extern void be_load_md5_lib(bvm *vm);
 extern void be_load_webclient_lib(bvm *vm);
 extern void be_load_crypto_lib(bvm *vm);
+extern void be_load_Leds_ntv_class(bvm *vm);
+extern void be_load_Leds_class(bvm *vm);
+extern void be_load_Leds_matrix_class(bvm *vm);
+extern void be_load_Leds_animator_class(bvm *vm);
 
 extern void be_load_ctypes_lib(bvm *vm);
 extern void be_load_ctypes_energy_definitions_lib(bvm *vm);
@@ -186,6 +190,12 @@ BERRY_API void be_load_custom_libs(bvm *vm)
 #if defined(USE_ONEWIRE) || defined(USE_DS18x20)
     be_load_onewirelib(vm);
 #endif
+#ifdef USE_WS2812
+    be_load_Leds_ntv_class(vm);
+    be_load_Leds_class(vm);
+    be_load_Leds_matrix_class(vm);
+    be_load_Leds_animator_class(vm);
+#endif // USE_WS2812
 #ifdef USE_I2S_AUDIO_BERRY
     be_load_driver_audio_lib(vm);
 #endif